buka: button::Main calls Deselect; mouse coords are i32
Retail's button::Main issues five calls to button::Deselect where ours
inlined the body (five UpdateScreenRegion expansions instead). Our
Deselect already existed as a one-line forwarder to an inline
DeselectSelected; Main just wasn't using it. 2015 -> 1179 bytes against
retail's 1240, diffs 1127 -> 926.

Same width fix as widget::Main applied to the mouse coordinates in
BUTTON and TEXTWDGT: msg.payload.mouse.x and m_owner->m_posX are both
i32, so narrowing either with static_cast<i16> emits WORD reads retail
does not have. textWidget::Main 733 -> 709 bytes, 499 -> 486 diffs.

Pattern worth naming, since it has now produced four closes and several
large reductions: where our reconstruction introduced a private inline
helper, retail frequently has a real method that already exists in our
class - widget::Dim, button::Deselect. The inline expansion is the
divergence; the method is sitting right there unused.

+++ src/BASE/TEXTWDGT.cpp
@@ -88,10 +88,8 @@ MessageDispatchResult textWidget::Main(tag_message& msg) {
     switch (msg.type) {
         case MESSAGE_LEFT_BUTTON_DOWN:
         case MESSAGE_RIGHT_BUTTON_DOWN: {
-            i16 relativeX =
-                static_cast<i16>(msg.payload.mouse.x) - static_cast<i16>(m_owner->m_posX);
-            i16 relativeY =
-                static_cast<i16>(msg.payload.mouse.y) - static_cast<i16>(m_owner->m_posY);
+            i32 relativeX = msg.payload.mouse.x - m_owner->m_posX;
+            i32 relativeY = msg.payload.mouse.y - m_owner->m_posY;
             if (relativeX < m_x || relativeY < m_y || relativeX >= m_x + m_width
                 || relativeY >= m_y + m_height)
                 return MESSAGE_DISPATCH_CONTINUE;
